do not delete parent directory after finish, if no files add mp4 and avi container fix cmd util rename templates update library getid3

+++ libs/mfr/Renamer.php
@@ -150,7 +150,7 @@ class Renamer extends \SA\Log\ClassWithErrorLog{
 			}
 		}
 		// check if cur dir is empty
-		if (!(new \FilesystemIterator($curFolder))->valid()) {
+		if (!(new \FilesystemIterator($curFolder))->valid() && ($aFolder !== '/')) {
 			$this->pput('Removing empty folder %s', $curFolder);
 			rmdir($curFolder);
 		}
